Fix for wrong order of adding the logging delegation handler
