+++ Fhi.HelseId.Blazor.Tests/LoggingDelegationHandlerTests.cs
@@ -1,8 +1,4 @@
-using Microsoft.Extensions.Logging;
-using Microsoft.Net.Http.Headers;
 using NUnit.Framework;
-using System;
-using System.Collections.Generic;
 using System.Linq;
 using System.Net.Http;
 using System.Threading.Tasks;
@@ -28,25 +24,5 @@ public class LoggingDelegationHandlerTests
         Assert.That(logger.Entries.Single(),
             Contains.Substring("with response 200 OK with CorrelationId TESTCORR"));
 
-    }
-}
-
-public class TestLogger<T> : ILogger<T>
-{
-    public List<string> Entries = [];
-
-    public IDisposable? BeginScope<TState>(TState state) where TState : notnull
-    {
-        return null;
-    }
-
-    public bool IsEnabled(LogLevel logLevel)
-    {
-        return true;
-    }
-
-    public void Log<TState>(LogLevel logLevel, EventId eventId, TState state, Exception? exception, Func<TState, Exception?, string> formatter)
-    {
-        Entries.Add(logLevel.ToString() + " " + formatter(state, exception));
     }
 }
